reduce rcpp functions
